requirements.txt
flask==2.0.3
mysql-connector==2.2.9
XlsxWriter==3.0.3
xlwt==1.3.0
pyexcel-xlsx==0.6.0
# pandas==1.1.5
requests==2.27.1
flask-unittest==0.1.3